Update patch set 1
Patch Set 1:

Terragrunt verify job was removed, but teraform verify job is still in effect. I can see in logs that it's failing at "terraform validate" operation:

Error: Unsupported argument
  on modules/kubernetes/kubernetes/main.tf line 16, in provider "kubernetes":
  16:   load_config_file = true

Can you comment on that? From my analysis it seems that the issue might be that terraform on each "init" operation get newest versions of provider plugins and current version of "openstack" is:

>>Installing terraform-providers/openstack v1.36.0<<

In past when the verison was 1.31 the verify job passed. Could you validate on your side with this provider plugin version?

Patch-set: 1
CC: Gerrit User 5097 <5097@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>